refactor(arch-review-phase-8a): split api-client god module by domain (R9)
Splits packages/web-next/lib/api-client.ts (1,491 LOC) into 22 domain
files at packages/web-next/lib/api/, each <= 200 LOC. Behavior-preserving
mechanical extraction; no functional changes.

- Round 1: core.ts (130 LOC) — HttpError, request, getApiBase (newly
  exported, was module-private), buildQueryString, ListEnvelope, type
  re-exports from ../types.
- Round 2: 21 parallel subagents wrote 21 domain files for captures,
  entities, briefs, stats, search, synthesize, intelligence, skills,
  commitments, settings, investments, voice, wiki, email, ingest,
  system-health, admin, mcp-activity, config, email-settings,
  service-health. Largest: ingest.ts (163 LOC).
- Round 3: lib/api/index.ts barrel + lib/api-client.ts rewritten as
  thin export * from './api' re-export. Zero consumer changes
  required across 84 import sites.

Bonus cleanup folded into extraction: inline import('./types').X
dynamic type imports in entities.ts + briefs.ts converted to static
top-of-file imports.

TanStack Query hook extraction deferred — file split is independently
valuable; hook extraction is design work and is being tracked as a
follow-up action item (A128).

Build: PASS. Lint clean vs A127 baseline. Shared build PASS.

Closes Phase 8a. Adds A128 (TQ hook extraction follow-up). Phase 8b
next: tab-split god pages + delete packages/web.
