+++ packages/web-next/lib/api/captures.ts
@@ -1,0 +1,52 @@
+/**
+ * Captures API namespace. Extracted from `lib/api-client.ts` (Phase 8a Round 2).
+ *
+ * Exports: `capturesApi`, `CapturesListParams`, `CreateCapturePayload`
+ */
+
+import type { Capture, CaptureType, CaptureSource, BrainView } from '../types'
+import { request, buildQueryString } from './core'
+import type { ListEnvelope } from './core'
+
+// ---------------------------------------------------------------------------
+// capturesApi
+// ---------------------------------------------------------------------------
+
+export interface CapturesListParams {
+  limit?: number
+  offset?: number
+  brain_view?: BrainView
+  capture_type?: CaptureType
+  source?: CaptureSource
+  pipeline_status?: string
+  /** Filter by `source_metadata.source_provider` — used by the Financial page provider tabs. */
+  source_provider?: string
+}
+
+export interface CreateCapturePayload {
+  content: string
+  capture_type: CaptureType
+  brain_view: BrainView
+  source?: CaptureSource
+}
+
+export const capturesApi = {
+  /** GET /api/v1/captures — paginated list */
+  list: (params: CapturesListParams = {}): Promise<ListEnvelope<Capture>> => {
+    const qs = buildQueryString(params)
+    return request<ListEnvelope<Capture>>(`/captures${qs}`)
+  },
+
+  /** GET /api/v1/captures/:id — single capture */
+  get: (id: string): Promise<Capture> => {
+    return request<Capture>(`/captures/${encodeURIComponent(id)}`)
+  },
+
+  /** POST /api/v1/captures — create a new capture */
+  create: (payload: CreateCapturePayload): Promise<{ id: string; pipeline_status: string; created_at: string }> => {
+    return request<{ id: string; pipeline_status: string; created_at: string }>('/captures', {
+      method: 'POST',
+      body: JSON.stringify(payload),
+    })
+  },
+}
