+++ packages/web-next/lib/api/core.ts
@@ -1,0 +1,153 @@
+/**
+ * Shared utilities for the split-by-domain API client. Imported by every
+ * `lib/api/<domain>.ts` file. Re-exports types from `../types` for the
+ * public barrel so consumers can `import { ... } from '@/lib/api-client'`.
+ */
+
+// ---------------------------------------------------------------------------
+// Type imports from the local types file (never from @open-brain/shared)
+// ---------------------------------------------------------------------------
+
+import type {
+  Capture,
+  CaptureType,
+  CaptureSource,
+  BrainView,
+  Entity,
+  EntityDetail,
+  EntityType,
+  Brief,
+  BriefDetail,
+  SearchResult,
+  DashboardStats,
+  MentionsTimelineResponse,
+  AskEntityResponse,
+  BoardCommitment,
+  CommitmentStatus,
+  Integration,
+  IntegrationStatus,
+  SettingEntry as SettingEntryType,
+  Trigger,
+  AIRoutingConfig,
+  ModelRoutingEntry,
+  EmailConfig,
+  EmailChannel,
+  EmailChannelStatus,
+  ServiceHealthStatus,
+} from '../types'
+
+// ---------------------------------------------------------------------------
+// Re-exported union types — callers can import from here as a convenience
+// ---------------------------------------------------------------------------
+
+export type { Capture, CaptureType, CaptureSource, BrainView, Entity, EntityDetail, EntityType, Brief, BriefDetail, SearchResult, DashboardStats, MentionsTimelineResponse, AskEntityResponse, BoardCommitment, CommitmentStatus, Integration, IntegrationStatus, Trigger, AIRoutingConfig, ModelRoutingEntry, EmailConfig, EmailChannel, EmailChannelStatus, ServiceHealthStatus }
+
+// ---------------------------------------------------------------------------
+// HttpError — thrown for all non-2xx responses
+// ---------------------------------------------------------------------------
+
+export class HttpError extends Error {
+  readonly status: number
+  readonly body: unknown
+  readonly path: string
+
+  constructor(status: number, body: unknown, path: string) {
+    super(`HTTP ${status} on ${path}`)
+    this.name = 'HttpError'
+    this.status = status
+    this.body = body
+    this.path = path
+  }
+}
+
+// ---------------------------------------------------------------------------
+// Core request wrapper
+// ---------------------------------------------------------------------------
+
+export function getApiBase(): string {
+  if (typeof window !== 'undefined') return '/api/v1'
+  const host = process.env.API_URL ?? process.env.NEXT_PUBLIC_API_URL ?? 'http://localhost:3002'
+  return `${host.replace(/\/$/, '')}/api/v1`
+}
+
+/**
+ * Low-level fetch wrapper. Prefixes `/api/v1`, injects `X-Open-Brain-Caller: web-ui`,
+ * sets `Content-Type: application/json` when a body is present, and throws
+ * `HttpError` on non-2xx responses.
+ */
+export async function request<T>(path: string, init: RequestInit = {}): Promise<T> {
+  const url = `${getApiBase()}${path}`
+
+  const headers: Record<string, string> = {
+    'X-Open-Brain-Caller': 'web-ui',
+    // Spread any caller-supplied headers so they can override defaults if needed.
+    ...(init.headers as Record<string, string> | undefined),
+  }
+
+  // Only set Content-Type for requests that carry a body.
+  if (init.body !== undefined && init.body !== null) {
+    if (!headers['Content-Type']) {
+      headers['Content-Type'] = 'application/json'
+    }
+  }
+
+  const response = await fetch(url, { ...init, headers })
+
+  if (!response.ok) {
+    let body: unknown
+    const contentType = response.headers.get('content-type') ?? ''
+    try {
+      body = contentType.includes('application/json')
+        ? await response.json()
+        : await response.text()
+    } catch {
+      body = null
+    }
+    throw new HttpError(response.status, body, path)
+  }
+
+  // 204 No Content — return undefined cast to T (callers should type as void)
+  if (response.status === 204) {
+    return undefined as unknown as T
+  }
+
+  return response.json() as Promise<T>
+}
+
+// ---------------------------------------------------------------------------
+// Query string builder
+// ---------------------------------------------------------------------------
+
+/**
+ * Build a `?key=value` query string. Skips `undefined` values. Arrays are
+ * serialized by repeating the key: `tags=a&tags=b`.
+ */
+export function buildQueryString(params: object): string {
+  const qs = new URLSearchParams()
+  for (const [key, value] of Object.entries(params as Record<string, unknown>)) {
+    if (value === undefined || value === null) continue
+    if (Array.isArray(value)) {
+      for (const item of value) {
+        if (item !== undefined && item !== null) {
+          qs.append(key, String(item))
+        }
+      }
+    } else {
+      qs.set(key, String(value))
+    }
+  }
+  const str = qs.toString()
+  return str ? `?${str}` : ''
+}
+
+// ---------------------------------------------------------------------------
+// Response envelope shapes (API-level, not UI-level)
+// ---------------------------------------------------------------------------
+
+/** Generic paginated list envelope returned by most list endpoints. */
+export interface ListEnvelope<T> {
+  items: T[]
+  total: number
+  limit: number
+  offset: number
+}
